// dashboard/src/utils/hasura-api/generated/schemas/customTypes.ts
/**
 * Generated by orval v8.4.0 🍺
 * Do not edit manually.
 * Hasura API
 * API for managing remote schemas and events in Hasura
 * OpenAPI spec version: 1.0.0
 */
import type { EnumTypeDefinition } from './enumTypeDefinition';
import type { InputObjectTypeDefinition } from './inputObjectTypeDefinition';
import type { ObjectTypeDefinition } from './objectTypeDefinition';
import type { ScalarTypeDefinition } from './scalarTypeDefinition';

/**
 * Custom GraphQL types used by action definitions
 */
export interface CustomTypes {
  /** Custom scalar type definitions */
  scalars?: ScalarTypeDefinition[];
  /** Custom enum type definitions */
  enums?: EnumTypeDefinition[];
  /** Custom input object type definitions */
  input_objects?: InputObjectTypeDefinition[];
  /** Custom object type definitions */
  objects?: ObjectTypeDefinition[];
}